Draft port to HIP
This commit changes a number of changes to port DGL to work on HIP.
There are also some mostly-unrelated QoL changes to the DGL build and
testing setup. This commit is a big dump that is a squash of various
other stuff so I can just get this out there. Properly staged and
documented changes to follow.

Hipification has been performed in place and the original files are
preserved with a ".prehip" extension. Those were added in a previous
commit to make this diff slightly smaller.